Test build to try and fix memory issues
Went through code and made sure objects were being nulled out and tried
disposing CPUCounter each time it's used seeing as PerformanceCounter
seems to supress the GC. Is probably a costly thing to do but
hey....this is why we test.

Thanks to the peoples of reddit for the help!